:construction: Debugging database

+++ src/main/java/com/schnabel/schnabel/pharmacies/repository/ITermRepository.java
@@ -3,5 +3,5 @@ package com.schnabel.schnabel.pharmacies.repository;
 import com.schnabel.schnabel.pharmacies.model.Term;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ITermRepository extends CrudRepository<Term, Integer> {
+public interface ITermRepository extends CrudRepository<Term, Long> {
 }
